corpus: the emission catches up -- 283 stdlib files and 15 behavioral goldens
The regeneration for the two converter roots, overlaid from a seeded reconvert.

A/B footprint: two temp roots seeded identically (src/core + version.props +
docs/validation, mirroring the src/ layout) and converted from that SAME seed by
the pre-fix and post-fix binaries, so the diff is this change alone and is
independent of any pre-existing corpus drift. Marker gate on BOTH roots, line-
anchored and path-precise: 41 marked / 15 with a .cs.auto sibling / 0 clobbered.

A control comparison of the PRE-fix root against the committed tree found zero
production .cs/.csproj/README.md content drift (52 raw differences, all CR-
stripped-equal CRLF phantoms; the only real ones were 12 stale .cs.auto, the
known standing family), so the committed tree sat exactly on the -stdlib
emission point and this overlay banks precisely what is classified below.

287 files differ; ZERO unclassified:

  1,213  CS0219 -- removed dead named-return declaration
    437  CS0219 -- the orphaned blank separator that followed the block
                   (every one in a file that also lost a declaration)
    165  CS8778 -- old (nint)/(nuint) cast removed
    157  CS8778 -- new bare `...L` (the int64 resolution)
      8  CS8778 -- new unchecked((n[u]int)...)

By artifact type the diff is .cs ONLY -- no .csproj, no README.md, no .slnx --
which is what an emission-only change must look like. 283 .cs overlaid; the 4
.cs.auto excluded per the standard overlay rule. Those four are the review
siblings of sync/atomic/type.cs and syscall/exec_windows.cs, the hand-owned
files that carry 33 of the residual CS0219 -- so the siblings DO hold the fixed
form, which is the tell that the residual is hand-own staleness (CleanupBacklog
item 18), not a converter gap.

Behavioral: 15 .cs re-transpiled and their 15 goldens re-baselined via
UpdateTestTargets --createTargetFiles after the suite's own transpile. One
golden per changed .cs and no *Tests.cs churn, confirming no project was added
or removed.

Corpus build, src/go2cs-stdlib.slnx, 304 projects, -c Debug -t:Rebuild,
isolated: 1,945 -> 158 warnings, 0 errors. CS8778 620 -> 0, CS0219 1,219 -> 52,
and every one of the other 21 codes moved by exactly 0.

+++ src/core/syscall/types_windows.cs
@@ -299,7 +299,7 @@ public static Timeval /*tv*/ NsecToTimeval(int64 nsec) {
     // 100-nanosecond intervals since January 1, 1601
     var nsec = ((int64)ft.HighDateTime << (int)(32)) + (int64)ft.LowDateTime;
     // change starting time to the Epoch (00:00:00 UTC, January 1, 1970)
-    nsec -= (nint)116444736000000000L;
+    nsec -= 116444736000000000L;
     // convert into nanoseconds
     nsec *= 100;
     return nsec;
@@ -311,10 +311,10 @@ public static Filetime /*ft*/ NsecToFiletime(int64 nsec) {
     // convert into 100-nanosecond
     nsec /= 100;
     // change starting time to January 1, 1601
-    nsec += (nint)116444736000000000L;
+    nsec += 116444736000000000L;
     // split into high / low
-    ft.LowDateTime = (uint32)((int64)(nsec & (nint)0xffffffffL));
-    ft.HighDateTime = (uint32)((int64)((nsec >> (int)(32)) & (nint)0xffffffffL));
+    ft.LowDateTime = (uint32)((int64)(nsec & 0xffffffffL));
+    ft.HighDateTime = (uint32)((int64)((nsec >> (int)(32)) & 0xffffffffL));
     return ft;
 }
